# order-manager/contracts/order-manager.openapi.yaml
openapi: 3.1.0
info:
  title: Order Manager API
  version: 1.0.0
  description: API for furniture facade order management

servers:
  - url: http://localhost:3000/api/v1
    description: Local development

paths:
  /health:
    get:
      summary: Health check
      operationId: healthCheck
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  timestamp:
                    type: string
                    format: date-time
                  service:
                    type: string

  /version:
    get:
      summary: API version
      operationId: getVersion
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  version:
                    type: string
                  api:
                    type: string

  /orders:
    get:
      summary: List orders
      operationId: listOrders
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
    post:
      summary: Create order
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Bad request

  /orders/{orderId}:
    get:
      summary: Get order by ID
      operationId: getOrder
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Not found

components:
  schemas:
    Order:
      type: object
      properties:
        id:
          type: string
        customerId:
          type: string
        status:
          type: string
          enum: [draft, confirmed, in_production, shipped, delivered, cancelled]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
    OrderItem:
      type: object
      properties:
        id:
          type: string
        orderId:
          type: string
        sequenceAuto:
          type: integer
        sequenceManual:
          type: integer
          nullable: true
        elementType:
          type: string
        width:
          type: number
        height:
          type: number
        quantity:
          type: integer
        materialId:
          type: string
        coatingId:
          type: string
          nullable: true
        designId:
          type: string
          nullable: true
        decor:
          type: string
          nullable: true
        previewUrl:
          type: string
          nullable: true
    CreateOrderRequest:
      type: object
      required:
        - customerId
      properties:
        customerId:
          type: string
        status:
          type: string
          default: draft